Change ImageService to create random filename
Removes the requirement to have image DTOs since filename will be generated at server. Also do not need to sanitize filenames.

+++ src/main/java/mep/mvcsocial/imageservice/service/ImageServiceImpl.java
@@ -1,11 +1,13 @@
 package mep.mvcsocial.imageservice.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import mep.mvcsocial.imageservice.ImageProperties;
 import mep.mvcsocial.imageservice.domain.Image;
 import mep.mvcsocial.imageservice.domain.ImageRepo;
 import org.springframework.core.io.PathResource;
 import org.springframework.core.io.Resource;
+import org.springframework.http.MediaType;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.annotation.PostConstruct;
@@ -13,9 +15,11 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.UUID;
 
 
 @RequiredArgsConstructor
+@Slf4j
 public class ImageServiceImpl implements ImageService {
 
     private final ImageRepo imageRepo;
@@ -23,8 +27,9 @@ public class ImageServiceImpl implements ImageService {
     private Path root;
 
     @Override
-    public Image createImage(Image newImage, MultipartFile rawFile) {
-        Image databaseImage = imageRepo.save(newImage);
+    public Image createImage(String userId, MultipartFile rawFile) {
+        Image toSave = new Image(userId, getRandomFileName(rawFile.getContentType()));
+        Image databaseImage = imageRepo.save(toSave);
 
         // saveImage to server
         File file = new File(root.resolve(databaseImage.getFilename()).toUri());
@@ -37,9 +42,21 @@ public class ImageServiceImpl implements ImageService {
         return databaseImage;
     }
 
+    private String getRandomFileName(String contentType) {
+        switch(contentType) {
+            case MediaType.IMAGE_JPEG_VALUE:
+                return UUID.randomUUID().toString() + ".jpg";
+            case MediaType.IMAGE_PNG_VALUE:
+                return UUID.randomUUID().toString() + ".png";
+            default:
+                throw new StorageException("File is not an image.");
+        }
+    }
+
     @Override
     public Image getImage(String imageId) {
-        return imageRepo.findById(imageId).get();
+        return imageRepo.findById(imageId)
+                .orElseThrow(() -> new ImageNotFoundException("Image with id " + imageId + " not found"));
     }
 
     @Override
@@ -64,9 +81,11 @@ public class ImageServiceImpl implements ImageService {
         imageRepo
                 .findById(imageId)
                 .ifPresent( image -> {
-                    root.resolve(image.getFilename()).toFile().delete();
-
-                    imageRepo.delete(image);
+                    boolean isDeleted = root.resolve(image.getFilename()).toFile().delete();
+                    if(isDeleted)
+                        imageRepo.delete(image);
+                    else
+                        log.error("Failed to delete image " + imageId + " from database. Not found.");
                 });
     }
